♻️ refactor: mock 데이터 삭제 및 타입 분리
