+++ src/features/application/model/types.ts
@@ -9,6 +9,27 @@ export type Role =
   | "springboot"
   | "nodejs"
 
+export interface FormField {
+  label: string
+  question: string
+  answer: string
+  required?: boolean
+}
+
+export interface ApplicantFormData {
+  applicantId: string
+  chapter: string
+  role: Role
+  projectName: string
+  challengerName: string
+  challengerUniversity: string
+  commonFields: FormField[]
+  roleSection?: {
+    title: string
+    fields: FormField[]
+  }
+}
+
 export interface AssignmentCount {
   current: number
   total: number
